Types and Functions

Index

AMLPipelineBase.BaseFilters.Imputer

AMLPipelineBase.BaseFilters.OneHotEncoder

AMLPipelineBase.BaseFilters.Wrapper

AMLPipelineBase.BaseFilters.createtransformer

AMLPipelineBase.BaselineModels.Baseline

AMLPipelineBase.BaselineModels.Baseline

AMLPipelineBase.BaselineModels.Identity

AMLPipelineBase.BaselineModels.Identity

AMLPipelineBase.CrossValidators.crossvalidate

AMLPipelineBase.DecisionTreeLearners.Adaboost

AMLPipelineBase.DecisionTreeLearners.PrunedTree

AMLPipelineBase.DecisionTreeLearners.RandomForest

AMLPipelineBase.EnsembleMethods.BestLearner

AMLPipelineBase.EnsembleMethods.StackEnsemble

AMLPipelineBase.EnsembleMethods.VoteEnsemble

AMLPipelineBase.FeatureSelectors.CatFeatureSelector

AMLPipelineBase.FeatureSelectors.CatNumDiscriminator

AMLPipelineBase.FeatureSelectors.CatNumDiscriminator

AMLPipelineBase.FeatureSelectors.FeatureSelector

AMLPipelineBase.FeatureSelectors.FeatureSelector

AMLPipelineBase.FeatureSelectors.FeatureSelector

AMLPipelineBase.FeatureSelectors.NumFeatureSelector

TSML.MLBaseWrapper.StandardScaler

TSML.MLBaseWrapper.Standardize

TSML.Monotonicers.Monotonicer

TSML.Normalizers.Normalizer

TSML.Outliernicers.Outliernicer

AMLPipelineBase.Pipelines.ComboPipeline

AMLPipelineBase.Pipelines.Pipeline

AMLPipelineBase.Pipelines.Pipeline

AMLPipelineBase.Pipelines.Pipeline

TSML.Plotters.Plotter

TSML.Statifiers.Statifier

TSML.TSClassifiers.TSClassifier

TSML.ValDateFilters.CSVDateValReader

TSML.ValDateFilters.CSVDateValWriter

TSML.ValDateFilters.DateValLinearImputer

TSML.ValDateFilters.DateValMultiNNer

TSML.ValDateFilters.DateValNNer

TSML.ValDateFilters.DateValgator

TSML.ValDateFilters.DateValizer

TSML.ValDateFilters.Dateifier

TSML.ValDateFilters.Matrifier

TSML.ValDateFilters.Imputers.AbstractContext

TSML.ValDateFilters.Imputers.Chain

TSML.ValDateFilters.Imputers.Chain

TSML.ValDateFilters.Imputers.Context

TSML.ValDateFilters.Imputers.DropObs

TSML.ValDateFilters.Imputers.DropVars

TSML.ValDateFilters.Imputers.Fill

TSML.ValDateFilters.Imputers.ImputeError

TSML.ValDateFilters.Imputers.Imputor

TSML.ValDateFilters.Imputers.Interpolate

TSML.ValDateFilters.Imputers.LOCF

TSML.ValDateFilters.Imputers.NOCB

TSML.ValDateFilters.Imputers.WeightedContext

TSML.ValDateFilters.Imputers.chain

TSML.ValDateFilters.Imputers.chain!

TSML.ValDateFilters.Imputers.chain!

TSML.ValDateFilters.Imputers.dropobs

TSML.ValDateFilters.Imputers.dropvars

TSML.ValDateFilters.Imputers.fill

TSML.ValDateFilters.Imputers.impute

TSML.ValDateFilters.Imputers.impute

TSML.ValDateFilters.Imputers.impute!

TSML.ValDateFilters.Imputers.impute!

TSML.ValDateFilters.Imputers.impute!

TSML.ValDateFilters.Imputers.impute!

TSML.ValDateFilters.Imputers.impute!

TSML.ValDateFilters.Imputers.interp

TSML.ValDateFilters.Imputers.ismissing!

TSML.ValDateFilters.Imputers.locf

TSML.ValDateFilters.Imputers.nocb

TSML.ValDateFilters.Imputers.splitkwargs

TSML.ValDateFilters.Imputers.srs

TSML.SVMModels.SVMModel

TSML.SVMModels.SVMModel

TSML.SVMModels.SVMModel

Descriptions

AMLPipelineBase.BaseFilters.Imputer — Type

Imputer(
   Dict(
      # Imputation strategy.
      # Statistic that takes a vector such as mean or median.
      :strategy => mean
   )
)

Imputes NaN values from Float64 features.

Implements fit! and transform.

AMLPipelineBase.BaseFilters.OneHotEncoder — Type

OneHotEncoder(Dict(
   # Nominal columns
   :nominal_columns => Int[],

   # Nominal column values map. Key is column index, value is list of
   # possible values for that column.
   :nominal_column_values_map => Dict{Int,Any}()
))

Transforms myinstances with nominal features into one-hot form and coerces the instance matrix to be of element type Float64.

Implements fit! and transform.

AMLPipelineBase.BaseFilters.Wrapper — Type

Wrapper(
   default_args = Dict(
      :name => "ohe-wrapper",
      # Transformer to call.
      :transformer => OneHotEncoder(),
      # Transformer args.
      :transformer_args => Dict()
   )
)

Wraps around a transformer.

Implements fit! and transform.

AMLPipelineBase.BaseFilters.createtransformer — Function

createtransformer(prototype::Transformer, args=Dict())

Create transformer

prototype: prototype transformer to base new transformer on

options: additional options to override prototype's options

Returns: new transformer.

AMLPipelineBase.BaselineModels.Baseline — Type

Baseline(
   default_args = Dict(
       :name => "baseline",
      :output => :class,
      :strat => mode
   )
)

Baseline model that returns the mode during classification.

AMLPipelineBase.BaselineModels.Baseline — Method

Baseline(name::String,opt...)

Helper function

AMLPipelineBase.BaselineModels.Identity — Type

Identity(args=Dict())

Returns the input as output.

AMLPipelineBase.BaselineModels.Identity — Method

Baseline(name::String,opt...)

Helper function

AMLPipelineBase.AbsTypes.fit! — Method

fit!(bsl::Baseline,x::DataFrame,y::Vector)

Get the mode of the training data.

AMLPipelineBase.AbsTypes.fit! — Method

fit!(idy::Identity,x::DataFrame,y::Vector)

Does nothing.

AMLPipelineBase.AbsTypes.transform! — Method

transform!(bsl::Baseline,x::DataFrame)

Return the mode in classification.

AMLPipelineBase.AbsTypes.transform! — Method

transform!(idy::Identity,x::DataFrame)

Return the input as output.

AMLPipelineBase.CrossValidators.crossvalidate — Method

crossvalidate(pl::Machine,X::DataFrame,Y::Vector,pfunc::Function,kfolds=10)

Run K-fold crossvalidation where:

pfunc is a performance metric

X and Y are input and target

AMLPipelineBase.DecisionTreeLearners.Adaboost — Type

Adaboost(
  Dict(
    :output => :class,
    :num_iterations => 7
  )
)

Adaboosted decision tree stumps. See DecisionTree.jl's documentation

Hyperparameters:

:num_iterations => 7 (number of iterations of AdaBoost)

Implements fit!, transform!

AMLPipelineBase.DecisionTreeLearners.PrunedTree — Type

PrunedTree(
  Dict(
    :purity_threshold => 1.0,
    :max_depth => -1,
    :min_samples_leaf => 1,
    :min_samples_split => 2,
    :min_purity_increase => 0.0
  )
)

Decision tree classifier. See DecisionTree.jl's documentation

Hyperparmeters:

:purity_threshold => 1.0 (merge leaves having >=thresh combined purity)

:max_depth => -1 (maximum depth of the decision tree)

:min_samples_leaf => 1 (the minimum number of samples each leaf needs to have)

:min_samples_split => 2 (the minimum number of samples in needed for a split)

:min_purity_increase => 0.0 (minimum purity needed for a split)

Implements fit!, transform!

AMLPipelineBase.DecisionTreeLearners.RandomForest — Type

RandomForest(
  Dict(
    :output => :class,
    :num_subfeatures => 0,
    :num_trees => 10,
    :partial_sampling => 0.7,
    :max_depth => -1
  )
)

Random forest classification. See DecisionTree.jl's documentation

Hyperparmeters:

:num_subfeatures => 0 (number of features to consider at random per split)

:num_trees => 10 (number of trees to train)

:partial_sampling => 0.7 (fraction of samples to train each tree on)

:max_depth => -1 (maximum depth of the decision trees)

:min_samples_leaf => 1 (the minimum number of samples each leaf needs to have)

:min_samples_split => 2 (the minimum number of samples in needed for a split)

:min_purity_increase => 0.0 (minimum purity needed for a split)

Implements fit!, transform!

AMLPipelineBase.AbsTypes.fit! — Method

fit!(adaboost::Adaboost, features::DataFrame, labels::Vector)

Optimize the hyperparameters of Adaboost instance.

AMLPipelineBase.AbsTypes.fit! — Method

fit!(tree::PrunedTree, features::DataFrame, labels::Vector)

Optimize the hyperparameters of PrunedTree instance.

AMLPipelineBase.AbsTypes.fit! — Method

fit!(forest::RandomForest, features::T, labels::Vector) where {T<:Union{Vector,Matrix,DataFrame}}

Optimize the parameters of the RandomForest instance.

AMLPipelineBase.AbsTypes.transform! — Method

transform!(adaboost::Adaboost, features::T) where {T<:Union{Vector,Matrix,DataFrame}}

Predict using the optimized hyperparameters of the trained Adaboost instance.

AMLPipelineBase.AbsTypes.transform! — Method

transform!(ptree::PrunedTree, features::DataFrame)

Predict using the optimized hyperparameters of the trained PrunedTree instance.

AMLPipelineBase.AbsTypes.transform! — Method

transform!(forest::RandomForest, features::T) where {T<:Union{Vector,Matrix,DataFrame}}

Predict using the optimized hyperparameters of the trained RandomForest instance.

AMLPipelineBase.EnsembleMethods.BestLearner — Type

BestLearner(
   Dict(
      # Output to train against
      # (:class).
      :output => :class,
      # Function to return partitions of instance indices.
      :partition_generator => (instances, labels) -> kfold(size(instances, 1), 5),
      # Function that selects the best learner by index.
      # Arg learner_partition_scores is a (learner, partition) score matrix.
      :selection_function => (learner_partition_scores) -> findmax(mean(learner_partition_scores, dims=2))[2],      
      # Score type returned by score() using respective output.
      :score_type => Real,
      # Candidate learners.
      :learners => [PrunedTree(), Adaboost(), RandomForest()],
      # Options grid for learners, to search through by BestLearner.
      # Format is [learner_1_options, learner_2_options, ...]
      # where learner_options is same as a learner's options but
      # with a list of values instead of scalar.
      :learner_options_grid => nothing
   )
)

Selects best learner from the set by performing a grid search on learners if grid option is indicated.

AMLPipelineBase.EnsembleMethods.StackEnsemble — Type

StackEnsemble(
   Dict(    
      # Output to train against
      # (:class).
      :output => :class,
      # Set of learners that produce feature space for stacker.
      :learners => [PrunedTree(), Adaboost(), RandomForest()],
      # Machine learner that trains on set of learners' outputs.
      :stacker => RandomForest(),
      # Proportion of training set left to train stacker itself.
      :stacker_training_proportion => 0.3,
      # Provide original features on top of learner outputs to stacker.
      :keep_original_features => false
   )
)

An ensemble where a 'stack' of learners is used for training and prediction.

AMLPipelineBase.EnsembleMethods.VoteEnsemble — Type

VoteEnsemble(
   Dict( 
      # Output to train against
      # (:class).
      :output => :class,
      # Learners in voting committee.
      :learners => [PrunedTree(), Adaboost(), RandomForest()]
   )
)

Set of machine learners employing majority vote to decide prediction.

Implements: fit!, transform!

AMLPipelineBase.AbsTypes.fit! — Method

fit!(bls::BestLearner, instances::DataFrame, labels::Vector)

Training phase:

obtain learners as is if grid option is not present

generate learners if grid option is present

foreach prototype learner, generate learners with specific options found in grid

generate partitions

train each learner on each partition and obtain validation output

AMLPipelineBase.AbsTypes.fit! — Method

fit!(se::StackEnsemble, instances::DataFrame, labels::Vector)

Training phase of the stack of learners.

perform holdout to obtain indices for

partition learner and stacker training sets

partition training set for learners and stacker

train all learners

train stacker on learners' outputs

build final model from the trained learners

AMLPipelineBase.AbsTypes.fit! — Method

fit!(ve::VoteEnsemble, instances::DataFrame, labels::Vector)

Training phase of the ensemble.

AMLPipelineBase.AbsTypes.transform! — Method

transform!(bls::BestLearner, instances::DataFrame)

Choose the best learner based on cross-validation results and use it for prediction.

AMLPipelineBase.AbsTypes.transform! — Method

transform!(se::StackEnsemble, instances::DataFrame)

Build stacker instances and predict

AMLPipelineBase.AbsTypes.transform! — Method

transform!(ve::VoteEnsemble, instances::DataFrame)

Prediction phase of the ensemble.

AMLPipelineBase.FeatureSelectors.CatFeatureSelector — Type

CatFeatureSelector(Dict(:name => "catf"))

Automatically extract categorical columns based on inferred element types.

Implements fit! and transform!.

AMLPipelineBase.FeatureSelectors.CatNumDiscriminator — Type

CatNumDiscriminator(
   Dict(
      :name => "catnumdisc",
      :maxcategories => 24
   )
)

Transform numeric columns to string (as categories) if the count of their unique elements <= maxcategories.

Implements fit! and transform!.

AMLPipelineBase.FeatureSelectors.CatNumDiscriminator — Method

CatNumDiscriminator(maxcat::Int)

Helper function for CatNumDiscriminator.

AMLPipelineBase.FeatureSelectors.FeatureSelector — Type

FeatureSelector(
   Dict(
     :name => "featureselector",
     :columns => [col1, col2, ...]
   )
)

Returns a dataframe of the selected columns.

Implements fit! and transform!.

AMLPipelineBase.FeatureSelectors.FeatureSelector — Method

FeatureSelector(cols::Vector{Int})

Helper function for FeatureSelector.

AMLPipelineBase.FeatureSelectors.FeatureSelector — Method

FeatureSelector(cols::Vararg{Int})

Helper function for FeatureSelector.

AMLPipelineBase.FeatureSelectors.NumFeatureSelector — Type

NumFeatureSelector(Dict(:name=>"numfeatsel"))

Automatically extracts numeric features based on their inferred element types.

Implements fit! and transform!.

TSML.MLBaseWrapper.StandardScaler — Type

StandardScaler(
   Dict( 
      :impl_args => Dict(
          :center => true,
          :scale => true
      )
   )
)

Standardizes each feature using (X - mean) / stddev. Will produce NaN if standard deviation is zero.

TSML.MLBaseWrapper.Standardize — Type

Standardize(d::Int, m::Vector{Float64}, s::Vector{Float64})

Standardization type.

AMLPipelineBase.AbsTypes.fit! — Function

fit!(st::StandardScaler, features::T, labels::Vector=[]) where {T<:Union{Vector,Matrix,DataFrame}}

Compute the parameters to center and scale.

AMLPipelineBase.AbsTypes.transform! — Method

transform!(st::StandardScaler, features::T)  where {T<:Union{Vector,Matrix,DataFrame}}

Apply the computed parameters for centering and scaling to new data.

TSML.Monotonicers.Monotonicer — Type

Monotonicer()

Monotonic filter to detect and normalize two types of dataset:

daily monotonic

entirely non-decreasing/non-increasing data

Example:

fname = joinpath(dirname(pathof(TSML)),"../data/testdata.csv")
csvfilter = CSVDateValReader(Dict(:filename=>fname,:dateformat=>"dd/mm/yyyy HH:MM"))
valgator = DateValgator(Dict(:dateinterval=>Dates.Hour(1)))
valnner = DateValNNer(Dict(:dateinterval=>Dates.Hour(1)))
stfier = Statifier(Dict(:processmissing=>true))
mono = Monotonicer(Dict())

mypipeline = @pipeline csvfilter |> valgator |> mono |> stfier
result = fit_transform!(mypipeline)

Implements: fit!, transform!

AMLPipelineBase.AbsTypes.fit! — Function

fit!(st::Monotonicer,features::T, labels::Vector=[]) where {T<:Union{Vector,Matrix,DataFrame}}

A function that checks if features are two-column data of Dates and Values

AMLPipelineBase.AbsTypes.transform! — Method

transform!(st::Monotonicer, features::T) where {T<:Union{Vector,Matrix,DataFrame}}

Normalize monotonic or daily monotonic data by taking the diffs and counting the flips.

TSML.Normalizers.Normalizer — Type

Normalizer(Dict(
   :method => :zscore
))

Transforms continuous features into normalized form such as zscore, unitrange, square-root, log, pca, ppca with parameter:

:method => :zscore or :unitrange or :sqrt or :log or pca or ppca or fa

:zscore => standard z-score with centering and scaling

:unitrange => unit range normalization with centering and scaling

:sqrt => square-root transform

:pca => principal component analysis transform

:ppca => probabilistic pca

:fa => factor analysis

:log => log transform

Example:

function generatedf()
    Random.seed!(123)
    gdate = DateTime(2014,1,1):Dates.Minute(15):DateTime(2016,1,1)
    gval1 = rand(length(gdate))
    gval2 = rand(length(gdate))
    gval3 = rand(length(gdate))
    X = DataFrame(Date=gdate,Value1=gval1,Value2=gval2,Value3=gval3)
    X
end

X = generatedf()
norm = Normalizer(Dict(:method => :zscore))
fit!(norm,X)
res=transform!(norm,X)

Implements: fit!, transform!

AMLPipelineBase.AbsTypes.fit! — Function

fit!(st::Statifier, features::T, labels::Vector=[]) where {T<:Union{Vector,Matrix,DataFrame}}

Validate argument features other than dates are continuous.

AMLPipelineBase.AbsTypes.transform! — Method

transform!(norm::Normalizer, features::T) where {T<:Union{Vector,Matrix,DataFrame}}

Compute statistics.

TSML.Outliernicers.Outliernicer — Type

Outliernicer(Dict(
   :dateinterval => Dates.Hour(1),
   :nnsize => 1,
   :missdirection => :symmetric,
   :scale => 1.25
))

Detects outliers below or above (median-scaleiqr,median+scaleiqr) and calls DateValNNer to replace them with nearest neighbors.

Example:

fname = joinpath(dirname(pathof(TSML)),"../data/testdata.csv")
csvfilter = CSVDateValReader(Dict(:filename=>fname,:dateformat=>"dd/mm/yyyy HH:MM"))
valgator = DateValgator(Dict(:dateinterval=>Dates.Hour(1)))
valnner = DateValNNer(Dict(:dateinterval=>Dates.Hour(1)))
stfier = Statifier(Dict(:processmissing=>true))
mono = Monotonicer(Dict())
outliernicer = Outliernicer(Dict(:dateinterval=>Dates.Hour(1)))

mpipeline = @pipeline csvfilter |> valgator |> mono |> valnner |> outliernicer |> stfier
results = fit_transform!(mpipeline)

Implements: fit!, transform!

AMLPipelineBase.AbsTypes.fit! — Function

fit!(st::Outliernicer, features::T, labels::Vector=[]) where {T<:Union{Vector,Matrix,DataFrame}}

Check that features are two-colum data.

AMLPipelineBase.AbsTypes.transform! — Method

transform!(st::Outliernicer, features::T) where {T<:Union{Vector,Matrix,DataFrame}}

Locate outliers based on IQR factor and calls DateValNNer to replace them with nearest neighbors.

AMLPipelineBase.Pipelines.ComboPipeline — Type

ComboPipeline(machs::Vector{T}) where {T<:Machine}

Feature union pipeline which iteratively calls fit_transform of each element and concatenate their output into one dataframe.

Implements fit! and transform!.

AMLPipelineBase.Pipelines.Pipeline — Type

Pipeline(machs::Vector{<:Machine},args::Dict=Dict())

Linear pipeline which iteratively calls and passes the result of fit_transform to the succeeding elements in the pipeline.

Implements fit! and transform!.

AMLPipelineBase.Pipelines.Pipeline — Method

Pipeline(machs::Vector{<:Machine},args::Dict=Dict())

Helper function for Pipeline structure.

AMLPipelineBase.Pipelines.Pipeline — Method

Pipeline(machs::Vararg{Machine})

Helper function for Pipeline structure.

TSML.Plotters.Plotter — Type

Plotter( Dict( :interactive => false, :pdfoutput => false ) )

Plots a TS by default but performs interactive plotting if specified during instance creation.

:interactive => boolean to indicate whether to use interactive plotting with false as default

:pdfoutput => boolean to indicate whether ouput will be saved as pdf with false as default

Example:

csvfilter = CSVDateValReader(Dict(:filename=>fname,:dateformat=>"dd/mm/yyyy HH:MM")) pltr = Plotter(Dict(:interactive => false))

mpipeline = @pipeline csvfilter |> pltr myplot = fit_transform!(mpipeline)

Implements: fit!, transform!

AMLPipelineBase.AbsTypes.fit! — Function

fit!(pltr::Plotter, features::T, labels::Vector=[]) where {T<:Union{Vector,Matrix,DataFrame}}

Check validity of features: 2-column Date,Val data

AMLPipelineBase.AbsTypes.transform! — Method

transform!(pltr::Plotter, features::T) where {T<:Union{Vector,Matrix,DataFrame}}

Convert missing into NaN to allow plotting of discontinuities.

TSML.Statifiers.Statifier — Type

Statifier(Dict(
   :processmissing => true
))

Outputs summary statistics such as mean, median, quartile, entropy, kurtosis, skewness, etc. with parameter:

:processmissing => boolean to indicate whether to include missing data stats.

Example:

dt=[missing;rand(1:10,3);missing;missing;missing;rand(1:5,3)]
dat = DataFrame(Date= DateTime(2017,12,31,1):Dates.Hour(1):DateTime(2017,12,31,10) |> collect,
                Value = dt)

statfier = Statifier(Dict(:processmissing=>false))

fit!(statfier,dat)
results=transform!(statfier,dat)

Implements: fit!, transform!

AMLPipelineBase.AbsTypes.fit! — Function

fit!(st::Statifier, features::T=[], labels::Vector=[]) where {T<:Union{Vector,Matrix,DataFrame}}

Validate argument to make sure it's a 2-column format.

AMLPipelineBase.AbsTypes.transform! — Function

transform!(st::Statifier, features::T=[]) where {T<:Union{Vector,Matrix,DataFrame}}

Compute statistics.

TSML.TSClassifiers.TSClassifier — Type

TSClassifier(
   Dict(
      # training directory
      :trdirectory => "",
      :tstdirectory => "",
      :modeldirectory => "",
      :feature_range => 7:20,
      :juliarfmodelname => "juliarfmodel.serialized",
      # Output to train against
      # (:class).
      :output => :class,
      # Options specific to this implementation.
      :impl_args => Dict(
         # Merge leaves having >= purity_threshold CombineMLd purity.
         :purity_threshold => 1.0,
         # Maximum depth of the decision tree (default: no maximum).
         :max_depth => -1,
         # Minimum number of samples each leaf needs to have.
         :min_samples_leaf => 1,
         # Minimum number of samples in needed for a split.
         :min_samples_split => 2,
         # Minimum purity needed for a split.
         :min_purity_increase => 0.0
      )
   )
)

Given a bunch of time-series with specific types. Get the statistical features of each, use these as inputs to RF classifier with output as the TS type, train and test. Another option is to use these stat features for clustering and check cluster quality. If accuracy is poor, add more stat features and repeat same process as outlined for training and testing. Assume that each time-series is named based on their type which will be used as target output. For example, temperature time series will be named as temperature?.csv where ? is an integer. Loop over each file in a directory, get stat and record in a dictionary/dataframe, train/test. Default to using RandomForest for classification of data types.

AMLPipelineBase.AbsTypes.fit! — Function

fit!(tsc::TSClassifier, features::T=[], labels::Vector=[]) where {T<:Union{Vector,Matrix,DataFrame}}

Get the stats of each file, collect as dataframe, and train.

AMLPipelineBase.AbsTypes.transform! — Function

transform!(tsc::TSClassifier, features::T=[]) where {T<:Union{Vector,Matrix,DataFrame}}

Apply the learned parameters to the new data.

TSML.ValDateFilters.CSVDateValReader — Type

CSVDateValReader(
   Dict(
      :filename => "",
      :dateformat => ""
   )
)

Reads csv file and parse date using the given format.

:filename => complete path including filename of csv file

:dateformat => date format to parse

Example:

inputfile =joinpath(dirname(pathof(TSML)),"../data/testdata.csv")
csvreader = CSVDateValReader(Dict(:filename=>inputfile,:dateformat=>"d/m/y H:M"))
fit!(csvreader)
df = transform!(csvreader)

# using pipeline workflow
filter1 = DateValgator()
filter2 = DateValNNer(Dict(:nnsize=>1))
mypipeline = @pipeline csvreader |> filter1 |> filter2
fit!(mypipeline)
res=transform!(mypipeline)

Implements: fit!, transform!

TSML.ValDateFilters.CSVDateValWriter — Type

CSVDateValWriter(
   Dict(
      :filename => "",
      :dateformat => ""
   )
)

Writes the time series dataframe into a file with the given date format.

Example:

inputfile =joinpath(dirname(pathof(TSML)),"../data/testdata.csv")
outputfile = joinpath("/tmp/test.csv")
csvreader = CSVDateValReader(Dict(:filename=>inputfile,:dateformat=>"d/m/y H:M"))
csvwtr = CSVDateValWriter(Dict(:filename=>outputfile,:dateformat=>"d/m/y H:M"))
filter1 = DateValgator()
filter2 = DateValNNer(Dict(:nnsize=>1))
mypipeline = @pipeline csvreader |> filter1 |> filter2 |> csvwtr
res=fit_transform!(mypipeline)

# read back what was written to validate
csvreader = CSVDateValReader(Dict(:filename=>outputfile,:dateformat=>"y-m-d HH:MM:SS"))
fit!(csvreader)
transform!(csvreader)

Implements: fit!, transform!

TSML.ValDateFilters.DateValLinearImputer — Type

DateValLinearImputer(
   Dict(
      :dateinterval => Dates.Hour(1),
  )
)

Fills missings by linear interpolation.

:dateinterval => time period to use for grouping,

Example:

Random.seed!(123)
gdate = DateTime(2014,1,1):Dates.Minute(15):DateTime(2016,1,1)
gval = Array{Union{Missing,Float64}}(rand(length(gdate)))
gmissing = 50000
gndxmissing = Random.shuffle(1:length(gdate))[1:gmissing]
X = DataFrame(Date=gdate,Value=gval)
X.Value[gndxmissing] .= missing

dnnr = DateValLinearImputer()
fit!(dnnr,X)
transform!(dnnr,X)

Implements: fit!, transform!`

TSML.ValDateFilters.DateValMultiNNer — Type

DateValMultiNNer(
   Dict(
      :type => :knn # :linear
      :missdirection => :symmetric, #:reverse, # or :forward or :symmetric
      :dateinterval => Dates.Hour(1),
      :nnsize => 1,
      :strict => true,
      :aggregator => :median
  )
)

Fills missings with their nearest-neighbors. It assumes that first column is a Date class and the other columns are Union{Missings,Real}. It uses DateValNNer and DateValizer+Impute to process each numeric column concatendate with the Date column.

:type => type of imputation which can be a linear interpolation or nearest neighbor

:missdirection => direction to fill missing data (:symmetric, :reverse, :forward)

:dateinterval => time period to use for grouping,

:nnsize => neighborhood size,

:strict => boolean value to indicate whether to be strict about replacement or not,

`:aggregator => function to aggregate based on date interval

Example:

Random.seed!(123)
gdate = DateTime(2014,1,1):Dates.Minute(15):DateTime(2016,1,1)
gval1 = Array{Union{Missing,Float64}}(rand(length(gdate)))
gval2 = Array{Union{Missing,Float64}}(rand(length(gdate)))
gval3 = Array{Union{Missing,Float64}}(rand(length(gdate)))
gmissing = 50000
gndxmissing1 = Random.shuffle(1:length(gdate))[1:gmissing]
gndxmissing2 = Random.shuffle(1:length(gdate))[1:gmissing]
gndxmissing3 = Random.shuffle(1:length(gdate))[1:gmissing]
X = DataFrame(Date=gdate,Temperature=gval1,Humidity=gval2,Ozone=gval3)
X.Temperature[gndxmissing1] .= missing
X.Humidity[gndxmissing2] .= missing
X.Ozone[gndxmissing3] .= missing

dnnr = DateValMultiNNer(Dict(
      :type=>:linear,
      :dateinterval=>Dates.Hour(1),
      :nnsize=>10,
      :missdirection => :symmetric,
      :strict=>true,
      :aggregator => :mean))
fit!(dnnr,X)
transform!(dnnr,X)

Implements: fit!, transform!`

TSML.ValDateFilters.DateValNNer — Type

DateValNNer(
   Dict(
      :missdirection => :symmetric, #:reverse, # or :forward or :symmetric
      :dateinterval => Dates.Hour(1),
      :nnsize => 1,
      :strict => true,
      :aggregator => :median
  )
)

Fills missings with their nearest-neighbors.

:missdirection => direction to fill missing data (:symmetric, :reverse, :forward)

:dateinterval => time period to use for grouping,

:nnsize => neighborhood size,

:strict => boolean value to indicate whether to be strict about replacement or not,

`:aggregator => function to aggregate based on date interval

Example:

Random.seed!(123)
gdate = DateTime(2014,1,1):Dates.Minute(15):DateTime(2016,1,1)
gval = Array{Union{Missing,Float64}}(rand(length(gdate)))
gmissing = 50000
gndxmissing = Random.shuffle(1:length(gdate))[1:gmissing]
X = DataFrame(Date=gdate,Value=gval)
X.Value[gndxmissing] .= missing

dnnr = DateValNNer(Dict(
      :dateinterval=>Dates.Hour(1),
      :nnsize=>10,
      :missdirection => :symmetric,
      :strict=>true,
      :aggregator => :mean))
fit!(dnnr,X)
transform!(dnnr,X)

Implements: fit!, transform!`

TSML.ValDateFilters.DateValgator — Type

DateValgator(args=Dict())
   Dict(
    :dateinterval => Dates.Hour(1),
    :aggregator => :median
  )
)

Aggregates values based on date period specified.

Example:

# generate random values with missing data
Random.seed!(123)
gdate = DateTime(2014,1,1):Dates.Minute(15):DateTime(2016,1,1)
gval = Array{Union{Missing,Float64}}(rand(length(gdate)))
gmissing = 50000
gndxmissing = Random.shuffle(1:length(gdate))[1:gmissing]
X = DataFrame(Date=gdate,Value=gval)
X.Value[gndxmissing] .= missing

dtvlmean = DateValgator(Dict(
      :dateinterval=>Dates.Hour(1),
      :aggregator => :mean))
fit!(dtvlmean,X)
res = transform!(dtvlmean,X)

Implements: fit!, transform!

TSML.ValDateFilters.DateValizer — Type

DateValizer(
   Dict(
    :medians => DataFrame(),
    :dateinterval => Dates.Hour(1)
  )
)

Normalizes and cleans time series by replacing missings with global medians computed based on time period groupings.

Example:

# generate random values with missing data
Random.seed!(123)
gdate = DateTime(2014,1,1):Dates.Minute(15):DateTime(2016,1,1)
gval = Array{Union{Missing,Float64}}(rand(length(gdate)))
gmissing = 50000
gndxmissing = Random.shuffle(1:length(gdate))[1:gmissing]
X = DataFrame(Date=gdate,Value=gval)
X.Value[gndxmissing] .= missing

dvzr = DateValizer(Dict(:dateinterval=>Dates.Hour(1)))
fit!(dvzr,X)
transform!(dvzr,X)

Implements: fit!, transform!

TSML.ValDateFilters.Dateifier — Type

Dateifier(args=Dict())
   Dict(
    :ahead => 1,
    :size => 7,
    :stride => 1
   )
)

Converts a 1-D date series into sliding window matrix for ML training

Example:

dtr = Dateifier(Dict())
lower = DateTime(2017,1,1)
upper = DateTime(2018,1,31)
dat=lower:Dates.Day(1):upper |> collect
vals = rand(length(dat))
x=DataFrame(Date=dat,Value=vals)
fit!(dtr,x)
res = transform!(dtr,x)

Implements: 'fit!, transform!

TSML.ValDateFilters.Matrifier — Type

Matrifier(Dict(
   Dict(
    :ahead  => 1,
    :size   => 7,
    :stride => 1,
  )
)

Converts a 1-D timeseries into sliding window matrix for ML training:

:ahead => steps ahead to predict

:size => size of sliding window

:stride => amount of overlap in sliding window

Example:

mtr = Matrifier(Dict(:ahead=>24,:size=>24,:stride=>5))
lower = DateTime(2017,1,1)
upper = DateTime(2017,1,5)
dat=lower:Dates.Hour(1):upper |> collect
vals = 1:length(dat)
x = DataFrame(Date=dat,Value=vals)
fit!(mtr,x)
res = transform!(mtr,x)

Implements: fit!, transform

AMLPipelineBase.AbsTypes.fit! — Function

fit!(dnnr::DateValMultiNNer,xx::T,y::Vector=[]) where {T<:DataFrame}

Validates and checks arguments for errors.

AMLPipelineBase.AbsTypes.fit! — Function

fit!(dtr::Dateifier,xx::T,y::Vector=[]) where {T<:Union{Matrix,Vector,DataFrame}}

Computes range of dates to be used during transform.

AMLPipelineBase.AbsTypes.fit! — Function

fit!(mtr::Matrifier,xx::T,y::Vector=Vector()) where {T<:Union{Matrix,Vector,DataFrame}}

Checks and validate inputs are in correct structure

AMLPipelineBase.AbsTypes.fit! — Function

fit!(csvwtr::CSVDateValWriter,x::T=[],y::Vector=[]) where {T<:Union{DataFrame,Vector,Matrix}}

Makes sure filename and dateformat are not empty strings.

AMLPipelineBase.AbsTypes.fit! — Function

fit!(dnnr::DateValLinearImputer,xx::T,y::Vector=[]) where {T<:DataFrame}

Validates and checks arguments for errors.

AMLPipelineBase.AbsTypes.fit! — Function

fit!(dvmr::DateValgator,xx::T,y::Vector=[]) where {T<:Union{Matrix,DataFrame}}

Checks and validates arguments.

AMLPipelineBase.AbsTypes.fit! — Function

fit!(csvrdr::CSVDateValReader,x::T=[],y::Vector=[]) where {T<:Union{DataFrame,Vector,Matrix}}

Makes sure filename and dateformat are not empty strings.

AMLPipelineBase.AbsTypes.fit! — Function

fit!(dnnr::DateValNNer,xx::T,y::Vector=[]) where {T<:DataFrame}

Validates and checks arguments for errors.

AMLPipelineBase.AbsTypes.fit! — Function

fit!(dvzr::DateValizer,xx::T,y::Vector=[]) where {T<:DataFrame}

Validates input and computes global medians grouped by time period.

AMLPipelineBase.AbsTypes.transform! — Function

transform!(csvrdr::CSVDateValReader,x::T=[]) where {T<:Union{DataFrame,Vector,Matrix}}

Uses CSV package to read the csv file and converts it to dataframe.

AMLPipelineBase.AbsTypes.transform! — Method

transform!(csvwtr::CSVDateValWriter,x::T) where {T<:Union{DataFrame,Vector,Matrix}}

Uses CSV package to write the dataframe into a csv file.

AMLPipelineBase.AbsTypes.transform! — Method

transform!(dnnr::DateValLinearImputer,xx::T) where {T<:DataFrame}

Replaces missings by linear interpolation.

AMLPipelineBase.AbsTypes.transform! — Method

transform!(dnnr::DateValMultiNNer,xx::T) where {T<:DataFrame}

Replaces missings by nearest neighbor or linear interpolation by looping over the dataset for each column until all missing values are gone.

AMLPipelineBase.AbsTypes.transform! — Method

transform!(dnnr::DateValNNer,xx::T) where {T<:DataFrame}

Replaces missings by nearest neighbor looping over the dataset until all missing values are gone.

AMLPipelineBase.AbsTypes.transform! — Method

transform!(dvmr::DateValgator,xx::T) where {T<:DataFrame}

Aggregates values grouped by date-time period using aggregate function such as mean, median, maximum, minimum. Default is mean.

AMLPipelineBase.AbsTypes.transform! — Method

transform!(dvzr::DateValizer,xx::T) where {T<:DataFrame}

Replaces missing with the corresponding global medians with respect to time period.

AMLPipelineBase.AbsTypes.transform! — Method

transform!(dtr::Dateifier,xx::T) where {T<:Union{Matrix,Vector,DataFrame}}

Transforms to day of the month, day of the week, etc

AMLPipelineBase.AbsTypes.transform! — Method

transform!(mtr::Matrifier,xx::T) where {T<:Union{Matrix,Vector,DataFrame}}

Applies the parameters of sliding windows to create the corresponding matrix

TSML.ValDateFilters.Imputers.AbstractContext — Type

AbstractContext

An imputation context records summary information about missing data for an imputation algorithm. All AbstractContexts are callable with a function, which allows us to write code like:

context() do c
    # My imputation code using a clean context
end

This do-block will pass a fresh context to your code and apply the on_complete function on the resulting data and context state. By default, on_complete will throw an ImputeError if we have too many missing values.

TSML.ValDateFilters.Imputers.Chain — Type

Chain <: Imputor

Runs multiple Imputors on the same data in the order they're provided.

Fields

imputors::Array{Imputor}

TSML.ValDateFilters.Imputers.Chain — Method

Chain(imputors::Imputor...) -> Chain

Creates a Chain using the Imputors provided (ordering matters).

TSML.ValDateFilters.Imputers.Context — Method

Context

Records base information about the missing data and assume all observations are equally weighted.

Keyword Arguments

n::Int: number of observations

count::Int: number of missing values found

limit::Float64: portion of total values allowed to be imputed (should be between 0.0 and 1.0).

is_missing::Function: must return a Bool indicating if the value counts as missing

on_complete::Function: a function to run when imputation is complete

TSML.ValDateFilters.Imputers.DropObs — Type

DropObs(; context=Context)

Removes missing observations from the AbstractArray or Tables.table provided.

Keyword Arguments

context::AbstractContext=Context(): A context which keeps track of missing data summary information

Example

TSML.ValDateFilters.Imputers.DropVars — Type

DropVars(; context=Context())

Finds variables with too many missing values in a AbstractMatrix or Tables.table and removes them from the input data.

Keyword Arguments

context::AbstractContext: A context which keeps track of missing data summary information

Examples

TSML.ValDateFilters.Imputers.Fill — Type

Fill(; value=mean, context=Context())

Fills in the missing data with a specific value. The current implementation is univariate, so each variable in a table or matrix will be handled independently.

Keyword Arguments

value::Any: A scalar or a function that returns a scalar if passed the data with missing data removed (e.g, mean)

context::AbstractContext: A context which keeps track of missing data summary information

Example

TSML.ValDateFilters.Imputers.ImputeError — Type

ImputeError{T} <: Exception

Is thrown by impute methods when the limit of imputable values has been exceeded.

Fields

msg::T - the message to print.

TSML.ValDateFilters.Imputers.Imputor — Type

Imputor

An imputor stores information about imputing values in AbstractArrays and Tables.tables. New imputation methods are expected to sutype Imputor and, at minimum, implement the impute!(imp::<MyImputor>, data::AbstractVector) method.

TSML.ValDateFilters.Imputers.Interpolate — Type

Interpolate(; context=Context())

Performs linear interpolation between the nearest values in an vector. The current implementation is univariate, so each variable in a table or matrix will be handled independently.

WARNING: Missing values at the head or tail of the array cannot be interpolated if there are no existing values on both sides. As a result, this method does not guarantee that all missing values will be imputed.

Keyword Arguments

context::AbstractContext: A context which keeps track of missing data summary information

Example

TSML.ValDateFilters.Imputers.LOCF — Type

LOCF(; context=Context())

Last observation carried forward (LOCF) iterates forwards through the data and fills missing data with the last existing observation. The current implementation is univariate, so each variable in a table or matrix will be handled independently.

See also:

NOCB: Next Observation Carried Backward

WARNING: missing elements at the head of the array may not be imputed if there is no existing observation to carry forward. As a result, this method does not guarantee that all missing values will be imputed.

Keyword Arguments

context::AbstractContext: A context which keeps track of missing data summary information

Example

TSML.ValDateFilters.Imputers.NOCB — Type

NOCB(; context=Context())

Next observation carried backward (NOCB) iterates backwards through the data and fills missing data with the next existing observation.

See also:

LOCF: Last Observation Carried Forward

WARNING: missing elements at the tail of the array may not be imputed if there is no existing observation to carry backward. As a result, this method does not guarantee that all missing values will be imputed.

Keyword Arguments

context::AbstractContext: A context which keeps track of missing data summary information

Example

TSML.ValDateFilters.Imputers.WeightedContext — Method

WeightedContext(wv; limit=1.0, is_missing=ismissing, on_complete=complete)

Records information about the missing data relative to a set of weights. This context type can be useful if some missing observation are more important than others (e.g., more recent observations in time series datasets)

Arguments

wv::AbstractWeights: a set of statistical weights to use when evaluating the importance of each observation. Will be accumulated during imputation.

Keyword Arguments

num::Int: number of observations

s::Float64: sum of the weights of missing values

limit::Float64: portion of total values allowed to be imputed (should be between 0.0 and 1.0).

is_missing::Function: returns a Bool if the value counts as missing

on_complete::Function: a function to run when imputation is complete

Base.:∘ — Method

Compose new Imputor chains with the composition operator

Example

Base.findfirst — Method

findfirst(ctx::AbstractContext, data::AbstractVector) -> Int

Returns the first non-missing index in data.

Arguments

ctx::AbstractContext: the context to pass into ismissing!

data::AbstractVector: the data array to search

Returns

Int: the first index in data that isn't missing

Base.findlast — Method

findlast(ctx::AbstractContext, data::AbstractVector) -> Int

Returns the last non-missing index in data.

Arguments

ctx::AbstractContext: the context to pass into ismissing!

data::AbstractVector: the data array to search

Returns

Int: the last index in data that isn't missing

Base.findnext — Method

findnext(ctx::AbstractContext, data::AbstractVector) -> Int

Returns the next non-missing index in data.

Arguments

ctx::AbstractContext: the context to pass into ismissing!

data::AbstractVector: the data array to search

Returns

Int: the next index in data that isn't missing

TSML.ValDateFilters.Imputers.chain! — Method

chain!(data, missing::Function, imputors::Imputor...; kwargs...)

Creates a Chain with imputors and calls impute!(imputor, missing, data; kwargs...)

TSML.ValDateFilters.Imputers.chain! — Method

chain!(data, imputors::Imputor...; kwargs...)

Creates a Chain with imputors and calls impute!(data, imputor)

TSML.ValDateFilters.Imputers.chain — Method

chain(data, args...; kwargs...)

Copies the data before calling chain!(data, args...; kwargs...)

TSML.ValDateFilters.Imputers.dropobs — Function

Impute.dropobs(data; dims=1, context=Context())

Removes missing observations from the AbstractArray or Tables.table provided. See DropObs for details.

Example

julia> using DataFrames; using Impute: Impute, Context

julia> df = DataFrame(:a => [1.0, 2.0, missing, missing, 5.0], :b => [1.1, 2.2, 3.3, missing, 5.5])
5×2 DataFrames.DataFrame
│ Row │ a        │ b        │
│     │ Float64⍰ │ Float64⍰ │
├─────┼──────────┼──────────┤
│ 1   │ 1.0      │ 1.1      │
│ 2   │ 2.0      │ 2.2      │
│ 3   │ missing  │ 3.3      │
│ 4   │ missing  │ missing  │
│ 5   │ 5.0      │ 5.5      │

julia> Impute.dropobs(df; dims=2, context=Context(; limit=1.0))
3×2 DataFrames.DataFrame
│ Row │ a       │ b       │
│     │ Float64 │ Float64 │
├─────┼─────────┼─────────┤
│ 1   │ 1.0     │ 1.1     │
│ 2   │ 2.0     │ 2.2     │
│ 3   │ 5.0     │ 5.5     │

TSML.ValDateFilters.Imputers.dropvars — Function

Impute.dropvars(data; dims=1, context=Context())

Finds variables with too many missing values in a AbstractMatrix or Tables.table and removes them from the input data. See DropVars for details.

Example

TSML.ValDateFilters.Imputers.fill — Function

Impute.fill(data; value=mean, dims=1, context=Context())

Fills in the missing data with a specific value. See Fill for details.

Example

TSML.ValDateFilters.Imputers.impute! — Method

impute!(data::AbstractMatrix, imp::Imputor; kwargs...)

Impute the data in a matrix by imputing the values one variable at a time; if this is not the desired behaviour custom imputor methods should overload this method.

Arguments

data::AbstractMatrix: the data to impute

imp::Imputor: the Imputor method to use

Keywords

dims: The dimension to impute along (e.g., observations dim)

Returns

AbstractMatrix: the input data with values imputed

Example

TSML.ValDateFilters.Imputers.impute! — Method

impute!(data, missing::Function, method::Symbol=:interp, args...; limit::Float64=0.1)

Creates the appropriate Imputor type and Context (using missing function) in order to call impute!(data, imputor::Imputor) with them.

Arguments

data: the datset containing missing elements we should impute.

missing::Function: the missing data function to use

method::Symbol: the imputation method to use (options: [:drop, :fill, :interp, :locf, :nocb])

args::Any...: any arguments you should pass to the Imputor constructor.

limit::Float64: missing data ratio limit/threshold (default: 0.1)

TSML.ValDateFilters.Imputers.impute! — Method

impute!(data, method::Symbol=:interp, args...; limit::Float64=0.1)

Looks up the Imputor type for the method, creates it and calls impute!(data, imputor::Imputor) with it.

Arguments

data: the datset containing missing elements we should impute.

method::Symbol: the imputation method to use (options: [:drop, :fill, :interp, :locf, :nocb])

args::Any...: any arguments you should pass to the Imputor constructor.

limit::Float64: missing data ratio limit/threshold (default: 0.1)

TSML.ValDateFilters.Imputers.impute! — Method

impute!(data, imp::Chain)

Runs the Imputors on the supplied data.

Arguments

imp::Chain: the chain to run

data: our data to impute

Returns

our imputed data

TSML.ValDateFilters.Imputers.impute! — Method

impute!(table, imp::Imputor)

Imputes the data in a table by imputing the values 1 column at a time; if this is not the desired behaviour custom imputor methods should overload this method.

Arguments

imp::Imputor: the Imputor method to use

table: the data to impute

Returns

the input data with values imputed

Example

TSML.ValDateFilters.Imputers.impute — Method

impute(data, imp::Imputor; kwargs...)

Returns a new copy of the data with the missing data imputed by the imputor imp.

Keywords

dims: The dimension to impute along (e.g., observations dim)

TSML.ValDateFilters.Imputers.impute — Method

impute(data, args...; kwargs...)

Copies the data before calling impute!(new_data, args...; kwargs...)

TSML.ValDateFilters.Imputers.interp — Function

Impute.interp(data; dims=1, context=Context())

Performs linear interpolation between the nearest values in an vector. See Interpolate for details.

Example

TSML.ValDateFilters.Imputers.ismissing! — Method

ismissing!(ctx::AbstractContext, x) -> Bool

Uses ctx.is_missing to determine if x is missing. If x is a NamedTuple or an AbstractArray then ismissing! will return true if ctx.is_missing returns true for any element. The ctx.count is increased whenever whenever we return true and if ctx.count / ctx.num exceeds our ctx.limit we throw an ImputeError

Arguments

ctx::Context: the contextual information about missing information.

x: the value to check (may be an single values, abstract array or row)

TSML.ValDateFilters.Imputers.locf — Function

Impute.locf(data; dims=1, context=Context())

Iterates forwards through the data and fills missing data with the last existing observation. See LOCF for details.

Example

TSML.ValDateFilters.Imputers.nocb — Function

Impute.nocb(data; dims=1, context=Context())

Iterates backwards through the data and fills missing data with the next existing observation. See LOCF for details.

Example

TSML.ValDateFilters.Imputers.splitkwargs — Method

splitkwargs(::Type{T}, kwargs...) where T <: Imputor -> (imp, rem)

Takes an Imputor type with kwargs and returns the constructed imputor and the unused kwargs which should be passed to the impute! call.

NOTE: This is used by utility methods with construct and imputor and call impute in 1 call.

TSML.ValDateFilters.Imputers.srs — Function

Impute.srs(data; rng=Random.GLOBAL_RNG, context=Context())

Simple Random Sampling (SRS) imputation is a method for imputing both continuous and categorical variables. Furthermore, it completes imputation while preserving the distributional properties of the variables (e.g., mean, standard deviation).

Example

TSML.SVMModels.SVMModel — Type

SVMModel(
  Dict(
    :name => "svm",
    :svmtype => "SVC"
  )
)

Wrapper for LIBSVM.jl for pipeline integration.

Implements fit! and transform!.

TSML.SVMModels.SVMModel — Method

SVMModel(svmt::String,opt=Dict())

Helper function

TSML.SVMModels.SVMModel — Method

SVMModel(svmt::String;opt...)

Helper function

AMLPipelineBase.AbsTypes.fit! — Function

fit!(svm::SVMModel, features::DataFrame, labels::Vector=[])

Train the model using SVM fit!

Arguments

svm::SVMModel: custom type

features::DataFrame: input

labels::Vector=[]:

AMLPipelineBase.AbsTypes.transform! — Method

transform!(svm::SVMModel, features::DataFrame)

Predict using svm model

Arguments

svm::SVMModel: custom type

features::DataFrame: input