Types and Functions

Types and Functions

Index

TSML.DecisionTreeLearners.Adaboost

TSML.DecisionTreeLearners.PrunedTree

TSML.DecisionTreeLearners.RandomForest

TSML.Outliernicers.Outliernicer

TSML.Plotters.Plotter

TSML.TSMLTypes.fit!

TSML.TSMLTypes.fit!

TSML.TSMLTypes.fit!

TSML.TSMLTypes.transform!

TSML.TSMLTypes.transform!

TSML.TSMLTypes.transform!

TSML.TSMLTypes.transform!

Descriptions

TSML.DecisionTreeLearners.Adaboost — Type.

Adaboost(
  Dict(
    :output => :class,
    :num_iterations => 7
  )
)

Adaboosted decision tree stumps. See DecisionTree.jl's documentation

Hyperparameters:

:num_iterations => 7 (number of iterations of AdaBoost)

Implements fit!, transform!

source

TSML.DecisionTreeLearners.PrunedTree — Type.

PrunedTree(
  Dict(
    :purity_threshold => 1.0,
    :max_depth => -1,
    :min_samples_leaf => 1,
    :min_samples_split => 2,
    :min_purity_increase => 0.0
  )
)

Decision tree classifier. See DecisionTree.jl's documentation

Hyperparmeters:

:purity_threshold => 1.0 (merge leaves having >=thresh combined purity)

:max_depth => -1 (maximum depth of the decision tree)

:min_samples_leaf => 1 (the minimum number of samples each leaf needs to have)

:min_samples_split => 2 (the minimum number of samples in needed for a split)

:min_purity_increase => 0.0 (minimum purity needed for a split)

Implements fit!, transform!

source

TSML.DecisionTreeLearners.RandomForest — Type.

RandomForest(
  Dict(
    :output => :class,
    :num_subfeatures => 0,
    :num_trees => 10,
    :partial_sampling => 0.7,
    :max_depth => -1
  )
)

Random forest classification. See DecisionTree.jl's documentation

Hyperparmeters:

:num_subfeatures => 0 (number of features to consider at random per split)

:num_trees => 10 (number of trees to train)

:partial_sampling => 0.7 (fraction of samples to train each tree on)

:max_depth => -1 (maximum depth of the decision trees)

:min_samples_leaf => 1 (the minimum number of samples each leaf needs to have)

:min_samples_split => 2 (the minimum number of samples in needed for a split)

:min_purity_increase => 0.0 (minimum purity needed for a split)

Implements fit!, transform!

source

TSML.TSMLTypes.fit! — Method.

fit!(adaboost::Adaboost, features::T, labels::Vector) where {T<:Union{Vector,Matrix,DataFrame}}

Function to optimize the hyperparameters of Adaboost instance.

source

TSML.TSMLTypes.fit! — Method.

fit!(tree::PrunedTree, features::T, labels::Vector) where {T<:Union{Vector,Matrix,DataFrame}}

Function to optimize the hyperparameters of PrunedTree instance.

source

TSML.TSMLTypes.fit! — Method.

fit!(forest::RandomForest, features::T, labels::Vector) where {T<:Union{Vector,Matrix,DataFrame}}

Function to optimize the parameters of the RandomForest instance.

source

TSML.TSMLTypes.transform! — Method.

transform!(adaboost::Adaboost, features::T) where {T<:Union{Vector,Matrix,DataFrame}}

Function to predict using the optimized hyperparameters of the trained Adaboost instance.

source

TSML.TSMLTypes.transform! — Method.

transform!(tree::PrundTree, features::T) where {T<:Union{Vector,Matrix,DataFrame}}

Function to predict using the optimized hyperparameters of the trained PrunedTree instance.

source

TSML.TSMLTypes.transform! — Method.

transform!(forest::RandomForest, features::T) where {T<:Union{Vector,Matrix,DataFrame}}

Function to predict using the optimized hyperparameters of the trained RandomForest instance.

source

TSML.Outliernicers.Outliernicer — Type.

Outliernicer(Dict())

Detects outliers below or above (q25-iqr,q75+iqr) and replace them with missing so that ValNNer can use nearest neighbors to replace the missings.

source

TSML.Plotters.Plotter — Type.

Plotter()

Plots a TS by default but performs interactive plotting if specified during instance creation.

source

TSML.TSMLTypes.transform! — Method.

Convert missing into NaN for plotting discontinuity

source