Types and Functions

Types and Functions

Creates an API wrapper for DecisionTrees for pipeline workflow.

Index

TSML.Outliernicers.Outliernicer

TSML.Plotters.Plotter

Descriptions

TSML.Outliernicers.Outliernicer — Type.

Outliernicer(Dict())

Detects outliers below or above (q25-iqr,q75+iqr) and replace them with missing so that ValNNer can use nearest neighbors to replace the missings.

source

TSML.Plotters.Plotter — Type.

Plotter()

Plots a TS by default but performs interactive plotting if specified during instance creation.

source

TSML.TSMLTypes.transform! — Method.

Convert missing into NaN for plotting discontinuity

source